refactor apps to match storage/catalogue patterns
