+++ .sbt_metadata/common.json
@@ -1,0 +1,6 @@
+{
+  "id" : "common",
+  "folder" : "common",
+  "dependencyIds" : [
+  ]
+}
