+++ .sbt_metadata/items_api.json
@@ -2,5 +2,6 @@
   "id" : "items_api",
   "folder" : "items_api",
   "dependencyIds" : [
+    "common"
   ]
 }
