+++ common/src/main/scala/uk/ac/wellcome/platform/stacks/common/http/config/models/HTTPServerConfig.scala
@@ -1,0 +1,7 @@
+package uk.ac.wellcome.platform.stacks.common.sierra.config.models
+
+case class HTTPServerConfig(
+  host: String,
+  port: Int,
+  externalBaseURL: String
+)
