+++ common/src/main/scala/uk/ac/wellcome/platform/stacks/common/sierra/config/builders/SierraApiConfigBuilder.scala
@@ -1,0 +1,16 @@
+package uk.ac.wellcome.platform.stacks.common.sierra.config.builders
+
+import com.typesafe.config.Config
+import uk.ac.wellcome.platform.stacks.common.sierra.config.models.SierraApiConfig
+
+object SierraApiConfigBuilder {
+  def buildSierraApiConfig(config: Config): SierraApiConfig = {
+    val user = config.required[String]("sierra.auth.user")
+    val pass = config.required[String]("sierra.auth.pass")
+
+    SierraApiConfig(
+      user = user,
+      pass = pass
+    )
+  }
+}
