+++ common/src/main/scala/uk/ac/wellcome/platform/stacks/common/sierra/config/models/SierraApiConfig.scala
@@ -1,0 +1,3 @@
+package uk.ac.wellcome.platform.stacks.common.sierra.config.models
+
+case class SierraApiConfig(user: String, pass: String)
