+++ common/src/main/scala/uk/ac/wellcome/platform/stacks/common/sierra/models/SierraPatron.scala
@@ -1,0 +1,5 @@
+package uk.ac.wellcome.platform.stacks.common.sierra.models
+
+case class SierraPatron(id: String,
+                        names: List[String] = Nil,
+                        emails: List[String] = Nil)
