+++ items_api/src/main/scala/uk/ac/wellcome/platform/stacks/status/api/Main.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.requests.api
+package uk.ac.wellcome.platform.status.api
 
 import java.net.URL
 
@@ -6,10 +6,10 @@ import akka.actor.ActorSystem
 import akka.stream.ActorMaterializer
 import com.typesafe.config.Config
 import uk.ac.wellcome.monitoring.typesafe.MetricsBuilder
-import uk.ac.wellcome.platform.requests.api.config.builders.{HTTPServerBuilder, SierraApiConfigBuilder}
 import uk.ac.wellcome.platform.stacks.common.sierra.config.builders.{HTTPServerBuilder, SierraApiConfigBuilder}
 import uk.ac.wellcome.platform.stacks.common.sierra.http.{HttpMetrics, WellcomeHttpApp}
-import uk.ac.wellcome.platform.stacks.common.sierra.services.SierraApi
+import uk.ac.wellcome.platform.stacks.common.sierra.services.{HttpSierraApi, SierraApi}
+import uk.ac.wellcome.platform.status.api.config.builders.{HTTPServerBuilder, SierraApiConfigBuilder}
 import uk.ac.wellcome.typesafe.WellcomeTypesafeApp
 import uk.ac.wellcome.typesafe.config.builders.AkkaBuilder
 
@@ -28,12 +28,12 @@ object Main extends WellcomeTypesafeApp {
 
     val sierraApiConfig = SierraApiConfigBuilder.buildSierraApiConfig(config)
 
-    val router: RequestsApi = new RequestsApi {
+    val router: StatusApi = new StatusApi {
       override implicit val ec: ExecutionContext = ecMain
       override implicit val sierraApi: SierraApi = new SierraApi(sierraApiConfig)
     }
 
-    val appName = "RequestsApi"
+    val appName = "StatusApi"
 
     new WellcomeHttpApp(
       routes = router.routes,
@@ -43,7 +43,7 @@ object Main extends WellcomeTypesafeApp {
       ),
       httpServerConfig = HTTPServerBuilder.buildHTTPServerConfig(config),
       contextURL = new URL(
-        "https://api.wellcomecollection.org/item-requests/v1/context.json"),
+        "https://api.wellcomecollection.org/item-status/v1/context.json"),
       appName = appName
     )
   }
