--- items_api/src/main/scala/uk/ac/wellcome/platform/status/api/config/models/SierraApiConfig.scala
@@ -1,3 +1,0 @@
-package uk.ac.wellcome.platform.status.api.config.models
-
-case class SierraApiConfig(user: String, pass: String)
