+++ requests_api/src/main/scala/uk/ac/wellcome/platform/requests/api/RequestsApi.scala
@@ -5,6 +5,9 @@ import grizzled.slf4j.Logging
 
 import scala.concurrent.ExecutionContext
 import uk.ac.wellcome.json.JsonUtil._
+import uk.ac.wellcome.platform.stacks.common.catalogue.services.CatalogueApi
+import uk.ac.wellcome.platform.stacks.common.sierra.models.SierraPatron
+import uk.ac.wellcome.platform.stacks.common.sierra.services.SierraApi
 
 case class Root(status: String = "ok")
 case class TestResponse(workId: String, itemId: String)
@@ -14,7 +17,7 @@ trait RequestsApi extends Logging {
   import de.heikoseeberger.akkahttpcirce.FailFastCirceSupport._
 
   implicit val ec: ExecutionContext
-  implicit val sierraApi: HttpSierraApi
+  implicit val sierraApi: SierraApi
 
   val routes: Route = concat(
     pathSingleSlash {
@@ -34,7 +37,7 @@ trait RequestsApi extends Logging {
       case (_, itemId) =>
         post {
           entity(as[SierraPatron]) { patron =>
-            val sierraItemNumber = HttpCatalogueApi.getItemINumber(itemId)
+            val sierraItemNumber = CatalogueApi.getItemINumber(itemId)
             val holdRequest = sierraApi.postPatronPlaceHold(patron.id.toString, sierraItemNumber.get)
             complete(holdRequest)
           }
