--- requests_api/src/main/scala/uk/ac/wellcome/platform/requests/api/config/models/HTTPServerConfig.scala
@@ -1,7 +1,0 @@
-package uk.ac.wellcome.platform.requests.api.config.models
-
-case class HTTPServerConfig(
-  host: String,
-  port: Int,
-  externalBaseURL: String
-)
